src/main/java/vitali/homeTrainig/unitTasting/junit/service/UserService.java
package vitali.homeTrainig.unitTasting.junit.service;

import vitali.homeTrainig.unitTasting.junit.dto.User;

import java.util.*;

public class UserService {

    private final List<User> users = new ArrayList<>();
    public List<User> getAll() {
        return users;
    }

    public void add(User ... users) {
        this.users.addAll(Arrays.asList(users));
    }

    public Optional<User> login(String username, String password) {
        return users.stream()
                .filter(user -> user.getUsername().equals(username))
                .filter(user -> user.getPassword().equals(password))
                .findFirst();
    }
}
